Use f64 instead of f32 in core and compose

+++ compose/src/ymlfmt.rs
@@ -47,8 +47,8 @@ pub struct YSignal {
     #[serde(default)]
     pub twos_complement: bool,
 
-    pub scale: Option<f32>,
-    pub offset: Option<f32>,
+    pub scale: Option<f64>,
+    pub offset: Option<f64>,
 
     pub unit: Option<String>,
